requirements.txt
flask
flask-sqlalchemy
werkzeug
bleach
python-slugify
Pillow
python-dotenv



